requirements.txt
Pillow~=9.4.0
requests~=2.28.1
yattag~=1.15.0
